+++ src/features/admin/routes/preference_routes.py
@@ -2,12 +2,15 @@ from typing import Optional
 
 from dependency_injector.wiring import inject, Provide
 from fastapi import APIRouter, Depends
+from starlette.authentication import UnauthenticatedUser
 
 from src.core.exceptions.not_found_exception import NotFoundException
 from src.features.admin.admin_container import AdminContainer
 from src.features.admin.services import PreferenceServiceContract
+from src.features.auth.services.auth_service_contract import AuthServiceContract
 from src.models.preference import PreferenceIn
 from src.models.response import APIResponse, UpdateResponse, DeleteResponse, CreateResponse
+from src.utils.oauth2_utils import oauth2_scheme
 
 preference_router = APIRouter(prefix="/preferences")
 
@@ -16,7 +19,14 @@ preference_router = APIRouter(prefix="/preferences")
 @inject
 async def days(
         preference_service: PreferenceServiceContract = Depends(Provide[AdminContainer.preference_service]),
+        token: str = Depends(oauth2_scheme),
+        auth_service: AuthServiceContract = Depends(Provide[AdminContainer.auth_service]),
 ):
+    user = await auth_service.get_current_user(token)
+
+    if not user:
+        raise UnauthenticatedUser
+
     types = await preference_service.get_days()
     return APIResponse(data=types)
 
@@ -26,7 +36,14 @@ async def days(
 async def create_preference(
         preference_in: PreferenceIn,
         preference_service: PreferenceServiceContract = Depends(Provide[AdminContainer.preference_service]),
+        token: str = Depends(oauth2_scheme),
+        auth_service: AuthServiceContract = Depends(Provide[AdminContainer.auth_service]),
 ):
+    user = await auth_service.get_current_user(token)
+
+    if not user:
+        raise UnauthenticatedUser
+
     preference = await preference_service.create_preference(preference_in)
     return CreateResponse(data=preference)
 
@@ -38,8 +55,15 @@ async def get_all_preferences(
         page: int = 1,
         page_size: int = 20,
         search_query: Optional[str] = None,
-        paginate: bool = True
+        paginate: bool = True,
+        token: str = Depends(oauth2_scheme),
+        auth_service: AuthServiceContract = Depends(Provide[AdminContainer.auth_service]),
 ):
+    user = await auth_service.get_current_user(token)
+
+    if not user:
+        raise UnauthenticatedUser
+
     preferences = await preference_service.get_preferences(page=page, page_size=page_size, paginate=paginate, search_query=search_query,)
     return APIResponse(data=preferences)
 
@@ -51,8 +75,15 @@ async def get_preferences_by_teacher_id(
         page: int = 1,
         page_size: int = 20,
         search_query: Optional[str] = None,
-        paginate: bool = True
+        paginate: bool = True,
+        token: str = Depends(oauth2_scheme),
+        auth_service: AuthServiceContract = Depends(Provide[AdminContainer.auth_service]),
 ):
+    user = await auth_service.get_current_user(token)
+
+    if not user:
+        raise UnauthenticatedUser
+
     preferences = await preference_service.get_preferences_by_teacher_id(page=page, teacher_id=teacher_id, page_size=page_size, paginate=paginate, search_query=search_query,)
     return APIResponse(data=preferences)
 
@@ -62,7 +93,14 @@ async def get_preferences_by_teacher_id(
 async def get_preference(
         id: int,
         preference_service: PreferenceServiceContract = Depends(Provide[AdminContainer.preference_service]),
+        token: str = Depends(oauth2_scheme),
+        auth_service: AuthServiceContract = Depends(Provide[AdminContainer.auth_service]),
 ):
+    user = await auth_service.get_current_user(token)
+
+    if not user:
+        raise UnauthenticatedUser
+
     preference = await preference_service.get_preference_by_id(id)
 
     if not preference:
@@ -77,7 +115,14 @@ async def update_preference(
         id: int,
         preference_in: PreferenceIn,
         preference_service: PreferenceServiceContract = Depends(Provide[AdminContainer.preference_service]),
+        token: str = Depends(oauth2_scheme),
+        auth_service: AuthServiceContract = Depends(Provide[AdminContainer.auth_service]),
 ):
+    user = await auth_service.get_current_user(token)
+
+    if not user:
+        raise UnauthenticatedUser
+
     preference = await preference_service.update_preference(id, preference_in)
 
     if not preference:
@@ -91,7 +136,14 @@ async def update_preference(
 async def delete_preference(
         id: int,
         preference_service: PreferenceServiceContract = Depends(Provide[AdminContainer.preference_service]),
+        token: str = Depends(oauth2_scheme),
+        auth_service: AuthServiceContract = Depends(Provide[AdminContainer.auth_service]),
 ):
+    user = await auth_service.get_current_user(token)
+
+    if not user:
+        raise UnauthenticatedUser
+
     is_deleted = await preference_service.delete_preference(id)
 
     if not is_deleted:
